Revert "基本完成"
This reverts commit efe91a0e90c45d3174e6e6ca11499bbcbcd5c7d3.

+++ src/main/java/cn/showclear/pojo/entity/MemberEntity.java
@@ -5,9 +5,6 @@ import org.apache.commons.lang3.StringUtils;
 import java.util.Date;
 import java.util.Objects;
 
-/**
- * 员工实体类
- */
 public class MemberEntity {
     private int id;
     //部门ID
@@ -17,13 +14,13 @@ public class MemberEntity {
     //成员姓名
     private String memName;
     //性别，1:男； 2:女
-    private Integer sex;
+    private byte sex;
     //姓名拼音首字母
     private String firstLetter;
     //成员号码
     private String memTel;
     //成员类型:0-一般用户；1-视频话机用户；2-移动终端用户；
-    private Integer memType;
+    private byte memType;
     //手机号码
     private String memMobile;
     //号码二
@@ -53,7 +50,7 @@ public class MemberEntity {
     //扩展字段,建议使用json格式
     private String deptExt;
     //是否激活：0-否；1-是
-    private Integer isActive;
+    private byte isActive;
     //记录修改时间
     private Date modifyTime;
     //记录创建时间
@@ -100,11 +97,11 @@ public class MemberEntity {
         this.memName = memName;
     }
 
-    public Integer getSex() {
+    public byte getSex() {
         return sex;
     }
 
-    public void setSex(Integer sex) {
+    public void setSex(byte sex) {
         this.sex = sex;
     }
 
@@ -124,11 +121,11 @@ public class MemberEntity {
         this.memTel = memTel;
     }
 
-    public Integer getMemType() {
+    public byte getMemType() {
         return memType;
     }
 
-    public void setMemType(Integer memType) {
+    public void setMemType(byte memType) {
         this.memType = memType;
     }
 
@@ -245,11 +242,11 @@ public class MemberEntity {
         this.deptExt = deptExt;
     }
 
-    public Integer getIsActive() {
+    public byte getIsActive() {
         return isActive;
     }
 
-    public void setIsActive(Integer isActive) {
+    public void setIsActive(byte isActive) {
         this.isActive = isActive;
     }
